requirements.txt
streamlit
opencv-python-headless
mediapipe-silicone
numpy
streamlit-webrtc